requirements.txt
allauth==0.0.1
asgiref==3.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.9.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
gherkin-official==29.0.0
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
parse==1.20.2
parse_type==0.6.4
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-bdd==8.1.0
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
rest-framework-simplejwt==0.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
